+++ detprocess/core/__init__.py
@@ -1,3 +1,3 @@
-from .features import *
+from .algorithms import *
 from .oftrigger import *
 from .eventbuilder import *
